Basic implementation of task templates using cookies.
Templates store the data: customer, project, task type, story, task story and
telework. They are stored in a cookie when the button 'Create template' is
pressed, and they are listed in the side bar to be used. The description of the
task is used as the name of the template.