Added a web installation wizard.
It will ask for the DB connection data, fill the configuration file in and
setup the database schema and initial files. It can't create the DB, so the
user has to create it and give proper permissions. User also has to give write
permissions on config directory to create the file.