[#164] Add CTRL + S shortcut for save tasks.
I had to change how the task fields save their content into the Store.

Until now they were saved when the focus changed, so it won't work if the focus
is inside a modified field when you press CTRL+S.

Different solutions had to be implemented for every field type.