Use multiselection combo box to assign users to projects.
The project management screen used a two-pane widget where you had to
drag and drop users to assign them to the project. This widget was not
intuitive nor very productive, and had issues in some browsers, so I
am now replacing it with a multiselection combo box.

The new implementation has some known issues:

* removeValuesFromStore parameter of the combo should be false, but it
  causes a bug when you add a new entry to a list of users and then
  remove an existing one. I worked around this issue forcing a reload
  of the availableUsersStore every time the 'Assign people' subwindow
  is opened, but it should not be necessary.

* I have noticed strange things happening when you edit user
  assignation for some project, then cancel and edit assignations on
  a different project. I worked this around forcing rejectChanges() on
  the assignedUsersStore.

* When you add enough users to the combo so there is no space in the
  subwindow, it should be resized. I was not able to implement this
  behaviour so I just did the window big enough so you won't likely
  need it.

Nonetheless, the change will hopefully reduce the complexity of this
part of the code and have less bugs. Additionally, I have renamed some
variables to make code more legible.