[#133] Always open project details in a new window.
Double-clicking on a project from a report is often done to check some
details of a project, but that doesn't mean you want to close the
report on screen.

This solution is simple and prevents losing time re-entering report
filtering data again, at the cost of having to close a tab.