[#128] Reimplement date range constraints in DateIntervalForm
Instead of using the validation mechanism to implement the date range
constraints, we implement it in the change event.

Doing that, we work around the problem that caused the bug #128: the
allowBlank attribute set to true directly marked the field as valid
when setting it empty, preventing our custom validation code, which
implemented the range constraints, to be run.