Added quick-access buttons to DateIntervalForm.
The new buttons allow to quickly get a report for the current or
previous weeks, or the current month.

The layout of the DateIntervalForm has changed. Now it does not extend
FormPanel, but Panel. A bottom bar has been added to the Panel, which
contains the new buttons, and the old elements of the panel are
contained in a sub-panel with 'form' layout, so the outcome is really
similar to what it was before.