[#231] Fix bug setting the proper field names in the project creation service
When the creation operation finishes, the store reads the response and
overwrites any local fields with the ones from the server. This is intended
mostly to get the project id which was empty on the client side.

The problem was that the project creation service returned the fields "initDate"
and "endDate" instead of "init" and "end", which were the names expected by the
client side. The client overwrote the local copy of the project with the empty
(because they were absent from the response) "init" and "end" properties.