[#233] Add week number to some calendar widgets
Replaced standard datefields with datefieldplus in the missing reports
(user tasks and project evaluation).

Quicktips have been explicitly disabled because it caused an exception
at line 1551 in ext.ux.datepickerplus.js when opening the calendar.
This is likely a bug in datepickerplus, we are just working it around.

Caught a typo in datepickerplus code and fixed it.