Prepare installation scripts for the upcoming 2.16 release.
This includes adding the DB changes for 2.16 to the installation
scripts.

Besides, corrected a mistake in the manual installation steps of the
admin docs and in the installation scripts themselves; the file
schema.sql should only contain the original 2.0 database, any later
changes to the DB are added by the file update/all.sql.