Generate config.defaults instead of keeping it under version control.
This file is used in the manual installation process, and it was not
up to date wrt config.php which is the file used on a development
environment.

To prevent this file from getting outdated again, this patch just
generates config.defaults from config.php when a release is made.