Preliminary implementation of persistent templates creation.
This patch contains the creation service and the page to migrate the
cookie based templates to persistent templates.

FIXME: the creation service contains hard-coded DB queries, violating
the design of the model layer.

FIXME: the migration page does not check if the migration completed
successfully and doesn't remove the templates cookie.