[#297] Simplify query for tasks page with a returnExtendedInfo flag
After PR#305 queries to getProjects service became too slow. It was
because big queries were being run to satisfy the requirements of the
project evaluation report.

Now the slower query is hidden behind a flag which enables the service
to retrieve dedication data from the project.

Collaterally fixed project evaluation report, the DAO was not
returning CustomProjectVO objects by mistake.