[#327] Add a template button to create a full-day task
The button adds one task with the init and end time set so it fills
the entire journey for the current date.

The JourneyHistoryDAO operation getByIntervals had to be exposed in
the user facade. It is called from tasks.php and saved to a javascript
variable so it can be used from tasks.js.