[#329] Use fullDescription as display field for the project combo
It simplifies the front-end code because it doesn't have to compose
the project description and customer name to display it, it just
reuses the value received from the server.

The function setCustomValue had to be renamed because its purpose has
changed now, but it is still useful to set the tooltip at the same
time the value of the combo box is set.