[#395] Clean up ext-all.js and related files.
There were several versions of the file in the repository, with
different degrees of minification. Remove all of them except for
the most verbose one.

Rename ext-all-debug-w-comments.js to simply ext.js. We chose a
different file name to make it easier to check the history of the
original file.

Pointed header.php to include ext.min.js as produced by the new
minify target in the makefile.