[#395] Clean up ext-base.js and related files.
There were minified and un-minified versions of the file in the
repository. Remove the minified version.

Rename ext-base-debug.js to simply base.js. We chose a
different file name to make it easier to check the history of the
original file.

Pointed header.php to include base.min.js as produced by the new
minify target in the makefile.