[#401] Remove explicit session id management in front-end.
Some inline PHP inside JS code took care of explicitly sending the
session id to web services. This is unnecessary, the browser already
sends the PHP session id via cookies.

This removes a bunch of ugly inline PHP, which is necessary to achieve
the separation of JS and PHP sources.