Fix side effect in task panel padding.
We added a CSS rule to set a 5px padding between task panels in commit
9abb57ffd9111102c3e5845d2566da6cfcc6e9f3.

This rule had a side effect, increasing the padding in sub-panels
inside the task panel.

We add a specific class to the top task panel container and modify the
rule to apply only to that content.