[#353] Fix wrong length of "full day task" on new users' first day.
On a new user's first day, $lastTaskDate is null and is initialized to
"the day before", this initialization had the side effect of modifying
the original $date object, so any date-related checks that happened
after that point were inadvertently using the previous day.

Additionally, changed behavior of default "full day" templates so they
don't create any task in case the journey for that date is zero.