[#431] Fix some 'undefined index' warnings in PostgreSQLTaskDAO.
Using is_null on an array element evaluates if the contents of that
position in the array is null, assuming the element exists, and emits
a warning if it doesn't.

Using isset checks the existence of that element, which is what we
actually want.

Fixes warnings like this, logged from many phpreport pages:
PHP Notice:  Undefined index:  in
/var/www/html/phpreport/model/dao/TaskDAO/PostgreSQLTaskDAO.php on line 409