[#401] Extract JS code from viewProjectDetails.php.
Original .php file was renamed to .js, to preserve history of the JS
code, which accounts for most of the file. The PHP code was moved to
a newly created .php file, some unused includes were removed but it's
otherwise unchanged.

We also take the chance to rename the report following current
conventions. New name is "projectDetailsReport".