[#472] Use correct journey value to calculate pending holidays.
In the case there were multiple contracts (journeys) between the start
of the year and the date in course, the code couldn't decide which one
to use and it resorted to 8 by default. To fix this, we made the query
more specific, setting the current date as both start and end values.

Also, do not use 8 as the fallback journey value so, in case the user
doesn't have a contract set in that moment, it reports values in HH:mm
format and doesn't attempt to convert it to days.