Implement server-side CSV output in the weekly project hours report.
This is controlled by a newly added GET parameter. Set "format=csv" for
it to output data in CSV format. Any other value for the format
parameter, or not setting it at all, will output the usual expected
JSON data.

We have changed the internal key used to identify the columns from
"$week-$year" to "$year-$week", so the list of columns can be easily
sorted. We also added an index key to $records as a way to sort the
rows. Sorting was not taken into consideration before, because it was
performed by the client side.

CSV export code is inspired by this example:
https://www.virendrachandak.com/techtalk/creating-csv-file-using-php-and-mysql/