Remove PartialUpdateReportAction.php.
The web service layer was not using this operation. Besides, it didn't
check overlapping between existing and the updated task, allowing
incorrect tasks to be saved.

We keep the entry point in TasksFacade, but running PartialUpdateReports
instead. Doing this, we only have to run permission and overlapping
checks in one place, in the PartialUpdateReports code path.

Also document the matter with overlapping checks in PostgreSQLTaskDAO.