Remove UpdateReportAction.php and associated facade operations.
This operation was not used by the web services layer and, in general,
it's not very useful because it updates all the fields of the task but
we almost never want to do that.

Besides, it was missing overlapping checks and also the new checks
related with project activation (#489), so it's dangerous to leave it
there and allow future developers to use it as is.

Add a WARNING to the corresponding DAO operation.