[#487][#491] Check lock on saved date before updating a task.
We were only checking the new task date in a task update operation,
instead of also checking the saved date. As a result, it was
possible to move a task from a locked date to an unlocked one, and
then update any other field.

As a side effect, we break the assumption in PartialUpdateTasksAction
that every DirtyTask contains a date, which helps fixing #491.