[#487] Break date assumption in checkOverlappingWithDBTasks.
Ensure that date field is extracted from the saved task in case the
DirtyTaskVO does not contain it.

Also use array_key_first() to get the first task in the list, because
a list of tasks where some were discarded might not start with 0. PHP
version in the requirements has been bumped to 7.3.0 due to this.