Fix regression in accumulated hours report with LDAP enabled.
This is a regression caused by commit
981533e62f1e83dbd68cd80dc215385c37d86c90.

The function LDAPBelongsDAO->setAValues may be called from
BaseRelationshipDAO->executeFromB, so it's not true that it never
receives the id and password fields.

Restored the original code in LDAPBelongsDAO->setAValues, and fixed
the PHP warning that we wanted to address in a different way.