Remove CustomEvent entity and DAO.
This was part of a feature from PhpReport 1.x, like ProjectSchedule.
It was never implemented after the rewrite.

Keeping this old code around will make our work to modernize the
backend more difficult, it's better to get rid of it.

We remove the entity, DAO, Actions and facade functions. The tables
are still in the database.