Consider partial leaves when grouping by week
Instead of counting as a full day every day that has a leave,
proportionate when there are partial leaves in the week.

This was causing some confusion because the sum was always
considering each day that had a leave as 1 and creating
unprecise numbers.

Note that one day is the user journey at the time the page was
rendered.

Fixes https://github.com/Igalia/phpreport/issues/543