[#506] Setup of tasks store outside of onReady event.
This allows fetching the tasks earlier, speeding up the time until
the page is fully interactive.

The patch looks a bit messy but it mainly just moves the code to setup
the TaskStore object and the associated HttpProxy outside of the
onReady event. We take the chance to to rename these variables more
appropriately, and to fix some wrong indentation.

The load callback of the store does need to be run after the page is
ready, so we wrap its code inside Ext.onReady().