Centralize use of VACATIONS_PROJECT configuration parameters.
Expose the function TaskDAO::getVacationsProjectId() through the Facade
layer and use it everywhere we need the VACATIONS_PROJECT parameter.

That function will use the newly added VACATIONS_PROJECT_ID parameter
and fall back to the old one, logging an error message. This will allow
a smooth transition in any deployment.