Copy result of legacy CSV exporter to clipboard.
Apparently, the data URLs generated by the legacy CSV exporter are not
working in any browser.

Instead of trying to open a data URL, a quick workaround is to copy
the contents of the CSV report to the clipboard. Users would be able
to paste it into a text file, or even directly into a spreadsheet.

Copying to the clipboard appears to be more reliable than what we have
now.