Remove timeout for new, empty task creation.
The new, empty task that appears when no task are saved for the day
feels too slow. The reason is: if the function is called before tasks
have been loaded, it sets a 1-second timeout before retrying, which
is too slow.

To fix the issue, we refactor the code a bit so the function is called
as soon as possible, exactly when task load is done.