Make sure createVacations return flat arrays
In PHP, array_diff returns a multidimensional array with the index
the value had in its parent array. For instance:

A diff between ['a', 'b', 'c'] and ['a', 'b'] will return {'2': 'c'}.

We don't want the index of the values, because when converting to JSON
it will be converted to an object in case the index doesn't start in 0.

Use array_values to make sure it returns a flat array.