Fix Vacation Summary report Used Hours column
Case the user was filtering the report for another year that was not
the current one, the Used Hours column would be wrong because it was
filtering the hours until the current day instead of the last day
of the report year. This would result in a range of dates that
spawns from Jan 1st of the filtered year until the current day.

We want to use the current date as end date only if we are filtering
for the current year.