[#172] Delete user assignment when deleting a project.
We add a DELETE CASCADE restriction to remove entries in the project_usr
table at the same time we delete the corresponding project.

Other restrictions to delete projects are still in place, for example,
you cannot delete a project that has assigned tasks due to other foreign
key constraints.

This being a change in the database, we bump its version too, and write
an upgrade script.