[#496, #513] Replace project partial update operation.
We had a specific operation in the backend to update only certain
fields of a project, but the UI was always sending the full project
data.

We think the ability of updating only some fields of the project is
not important, and we prefer to minimize the amount of code to
maintain. So we have:
* Switched the web service to the previously unused UpdateProjectAction
  and ProjectDAO::update.
* Modify ProjectDAO::update to detect specific errors and return an
  OperationResult, just like the partialUpdate operation did.
* Modify ProjectDAO::update to use PDO.
* Remove all code related to PartialUpdateProject.