[#173,#496] Report better errors from LDAP user create DAO.
The HybridUserDAO matches user data with the LDAP and keeps them in
sync with the local database, to maintain the relations with other
data tables.

We modify the create operation in this DAO to return OperationResult,
matching the new expectations of the upper layers.

We change the Hybrid DAO to extend the Postgres DAO, so we can reuse
the SQL operations. In this case, we call the create operation from
the Postgres DAO after we do the corresponding LDAP checks.