[#173,#496] Report better errors from LDAP user update DAO.
This operation did not do anything related with LDAP, so we simply run
the corresponding SQL operation. We don't modify the behavior of the
operation, so we add a comment about a detected problem.

In the web service layer, we add try/catch blocks for the
LDAPInvalidOperationException like the ones in the create and delete
user services, to prevent crashing on operations that aren't defined
in the LDAP backend (namely, assigning users to groups).