Display if project has hours logged in the future
Sometimes users can mistakenly log hours in the future and this
was causing confusion because the total hours vs the hours displayed
per user in the report didn't match.

With this patch, if a project has hours logged in the future a field
with the amount of hours will be displayed in red in the project
details report.