Add React frontend with OIDC auth
These changes add a React frontend that will be the start of new pages in the application. OIDC authentication has been added as a context in the React app so that the app will have the user/token information everywhere - this makes it fairly straightforward to hit the new Python API with a bearer token.