Fix belongs table and initial migration
It had only the group id as a primary key, what would cause exceptions
as soon as a record using the same group id was inserted. This
wasn't detected in production because the tables were already there
when the migration was created, but when starting a db from scratch
it doesn't work well.