Add Task validation to TaskForm
When the form loads it will fetch all tasks from the current day
and after adding a task it updates the task global state with the new
added task.

It also adds the validation to the TaskForm, now it check's if
the schema is valid and if the tasks are overlapping. if any of those
checks fail, an error message will appear.