Create RefreshSession High Order Component
The RefreshSession HOC adds 2 pieces of logic to the components:

1. it refreshes the session each 5 minutes, so if a token is expired
it will try to refresh after 5minutes.
2. it checks if the token is expired on the window focus, and if it is
expired it will try to refresh the token.

This component is basically a wrapper for those funcionalities and
it was made as a HOC instead of a hooks because it needed to be used
within `providers.tsx` but it needed to be under `SessionProvider` since
it uses the `useSession` hook.