Create createTemplate api call
This commit changes the structure for the apiClient dependecy injection,
it exports a makeCreateTemplate function that receives the apiClient as
prop and return the createTemplate function, I made this change because
I think this pattern is easier to follow through the application. Also I
intend to change the other api calls in future pr's.