Add create and get project allocation endpoints
The creation is a simple method but to get the allocation ready
for the fronted, they are grouped by user and weeks instead of
being returned raw as the DB model.

This is to avoid too many calculations in the frontend to display
the data in a timeline.

In this phase of the project there are still too many business rules
missing like what to do with overlapping allocations and overbooking,
but this will be done in a later stage.