db: Add fxn/trigger to calculate vacation/expected
This function will run against each journey_history when one is inserted or updated. It will use the dates in the range and the journey, along with company fte and yearly vacation hours, to determine how much vacation a user should have in each year in the journey and how many expected hours a user should work in the year.